Create Red
it's morning it's morning The color of the sky is red. Breath in the chest, eternal imprisonment. The blood inside the vein was imprisoned. Someone was on the gallows. Someone whose life was destroyed before the execution. He objected to polluted and toxic air. He protested the right that was stolen from him. He objected to this dark night of oppression, to these executions and massacres, and to the little support they gave him. He protested for the dead